pyproject.toml
[tool.poetry]
name = "vps-control-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.28.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
